# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Good Food Tracker API",
    "description": "The project aims to allow the users to take pictures and/or leave notes, ratings, comments about restaurants they visit\r\n\tin order to be able to reference it later when they try to pick were they want to go eat out or order from.",
    "contact": {
      "name": "Igor Ilić",
      "email": "github@igorilic.net"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8098",
      "description": "Development API base url"
    },
    {
      "url": "https://gft.igorilic.net/api",
      "description": "Production API base url"
    }
  ],
  "paths": {
...
    "/country": {
      "get": {
        "tags": [
          "Country"
        ],
        "summary": "Fetch a list of countries",
        "description": "Endpoint used for get a list of countries",
        "operationId": "9b4823deea0af7e0075376465e07cc05",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Value used to filter the countries",
            "required": false,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Pagination start offset",
            "required": false,
            "schema": {
              "type": "integer",
              "additionalProperties": false
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination end offset",
            "required": false,
            "schema": {
              "type": "integer",
              "additionalProperties": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully filtered countries",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CountryEntity"
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Country"
        ],
        "summary": "Add new country",
        "description": "Endpoint used for adding a new country",
        "operationId": "2b7487fe8a3b5ecc1afe3b81f8f80fb3",
        "requestBody": {
          "description": "Required parameters",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CountryEntity"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required arguments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/country/{countryID}": {
      "get": {
        "tags": [
          "Country"
        ],
        "summary": "Fetch a information about a specific country",
        "description": "Endpoint used for getting information about a specific country",
        "operationId": "00b4848e7c68b94bfce85a91ff23d9ea",
        "parameters": [
          {
            "name": "countryID",
            "in": "path",
            "description": "ID of a country",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/uuid_parameter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CountryEntity"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required arguments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Country"
        ],
        "summary": "Delete country",
        "description": "Endpoint used for deleting a country",
        "operationId": "cddc8735b2a77841fee97c210153c20a",
        "parameters": [
          {
            "name": "countryID",
            "in": "path",
            "description": "ID of a country being updated",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/uuid_parameter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CountryEntity"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required arguments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "404": {
            "description": "Country not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Country"
        ],
        "summary": "Update country",
        "description": "Endpoint used for updating country information",
        "operationId": "69646430a79ad2e443a3df9538d77022",
        "parameters": [
          {
            "name": "countryID",
            "in": "path",
            "description": "ID of a country being updated",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/uuid_parameter"
            }
          }
        ],
        "requestBody": {
          "description": "Required parameters",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CountryEntity"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required arguments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "404": {
            "description": "Country not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountryEntity": {
        "description": "CountryEntity Class",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "error_response": {
        "properties": {
          "class": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "field": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "uuid_parameter": {
        "description": "Swagger doc block",
        "type": "string",
        "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[1-5]{1}[0-9A-Fa-f]{3}-[ABab89]{1}[0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}